refactor: move registry interfaces from repository.go to interfaces.go
Move Repository, BlobStore, ManifestStore, ReferenceFetcher, ReferencePusher,
ReferrerLister, TagLister, and Mounter interfaces to registry/interfaces.go.
The repository.go file retains only the Tags() and Referrers() helper functions.

This is a pure refactor with no behavior change.
